requirements.txt
altair
pandas
streamlit
jyotishyamitra
datetime
timedelta
plotly.express
pyswisseph==2.8.0.post1